pyproject.toml
[tool.poetry]
name = "data-base-manager"
version = "0.1.0"
description = ""
authors = ["SilverAnton <cu4hawk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg = "^3.1.18"
requests = "^2.31.0"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
